requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blis==1.0.1
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cssselect==1.2.0
cymem==2.0.8
Cython==0.29.23
distro==1.9.0
dnspython==2.7.0
einops==0.8.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
filelock==3.16.1
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
keras==2.14.0
kornia==0.7.3
kornia_rs==0.1.5
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.0
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.52.0
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
parse==1.20.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.10.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==13.9.2
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spandrel==0.4.0
srsly==2.4.8
sympy==1.13.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.0
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchsde==0.2.6
torchvision==0.19.1
tqdm==4.66.5
trampoline==0.1.2
transformers==4.45.2
tweepy==4.14.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==1.26.20
w3lib==2.2.1
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wrapt==1.14.1
yarl==1.13.1
zipp==3.20.2
